+++ models/UserOld.js
@@ -1,13 +1,12 @@
-// Dependencies
-// =============================================================
+// Requiring bcrypt for password hashing. Using the bcrypt-nodejs version as 
+//the regular bcrypt module sometimes causes errors on Windows machines
+var bcrypt = require("bcrypt-nodejs");
+//
+// Creating our User model
+//Set it as export because we will need it required on the server
 module.exports = function (sequelize, DataTypes) {
-  // This may be confusing but here Sequelize (capital) references the standard library
-
-  // sequelize (lowercase) references our connection to the DB.
-  // var db = require("../../config/connection");
-
-  // Creates a "user" model that matches up with DB
   var User = sequelize.define("User", {
+    // The email cannot be null, and must be a proper email before creation
     email: {
       type: DataTypes.STRING,
       allowNull: false,
@@ -16,18 +15,12 @@ module.exports = function (sequelize, DataTypes) {
         isEmail: true
       }
     },
+    // The password cannot be null
     password: {
       type: DataTypes.STRING,
       allowNull: false
-    },
-    name: DataTypes.STRING,
-    created_at: DataTypes.DATE
-  },
-    {
-
-      freezeTableName: true
-
-    });
+    }
+  });
   // Creating a custom method for our User model. 
   //This will check if an unhashed password entered by the 
   //user can be compared to the hashed password stored in our database
@@ -39,20 +32,5 @@ module.exports = function (sequelize, DataTypes) {
   User.beforeCreate(function (user) {
     user.password = bcrypt.hashSync(user.password, bcrypt.genSaltSync(10), null);
   });
-
-  User.associate = function (models) {
-    // Associating User with Posts
-    // When an User is deleted, also delete any associated Posts
-    User.hasMany(models.Transaction, {
-      onDelete: "cascade"
-    });
-  };
-
-  //   {
-
-  //     freezeTableName: true
-
-  // });
-
   return User;
 };
